+++ Controllers/PostsController.cs
@@ -12,6 +12,7 @@ namespace DAW_social_platform.Controllers
     {
         private ApplicationDbContext db = new ApplicationDbContext();
         // GET: Posts
+        [Authorize(Roles = "User,Visitor,Admin")]
         public ActionResult Index()
         {
             if (TempData.ContainsKey("message"))
@@ -24,12 +25,22 @@ namespace DAW_social_platform.Controllers
             return View();
         }
 
+        [Authorize(Roles = "User,Visitor,Admin")]
         public ActionResult Show(int id)
         {
             Post post = db.Posts.Find(id);
+
+            ViewBag.afisareButoane = false;
+            if (post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                ViewBag.afisareButoane = true;
+            }
+            ViewBag.esteAdmin = User.IsInRole("Admin");
+            ViewBag.utilizatorCurent = User.Identity.GetUserId();
             return View(post);
         }
 
+        [Authorize(Roles = "User,Visitor,Admin")]
         [HttpPost]
         public ActionResult Show(Comment comment)
         {
@@ -59,6 +70,7 @@ namespace DAW_social_platform.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         public ActionResult New()
         {
             Post post = new Post();
@@ -66,6 +78,7 @@ namespace DAW_social_platform.Controllers
             return View();
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public ActionResult New(Post post)
         {
@@ -91,27 +104,58 @@ namespace DAW_social_platform.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Edit(int id)
         {
             Post post = db.Posts.Find(id);
+            post.Date = DateTime.Now;
+            post.UserId = User.Identity.GetUserId();
+
+            if (post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                return View(post);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unei postari care nu va apartine";
+                return RedirectToAction("Index");
+            }
+
             return View(post);
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPut]
         public ActionResult Edit(int id, Post requestPost)
         {
             try
             {
-                Post post = db.Posts.Find(id);
-
-                if (TryUpdateModel(post))
+                if (ModelState.IsValid)
                 {
-                    post.Content = requestPost.Content;
-                    db.SaveChanges();
-                    TempData["message"] = "Postarea a fost modificata!";
-                    return RedirectToAction("Index");
+                    Post post = db.Posts.Find(id);
+
+                    if (post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+                    {
+                        if (TryUpdateModel(post))
+                        {
+                            post.Content = requestPost.Content;
+                            post.Date = DateTime.Now;
+                            post.UserId = User.Identity.GetUserId();
+                            db.SaveChanges();
+                            TempData["message"] = "Postarea a fost modificata!";
+                        }
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra uneri postari care nu va apartine";
+                        return RedirectToAction("Index");
+                    }
                 }
-                return View(requestPost);
+                else
+                {
+                    return View(requestPost);
+                }
             }
             catch (Exception e)
             {
@@ -119,14 +163,23 @@ namespace DAW_social_platform.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpDelete]
         public ActionResult Delete(int id)
         {
             Post post = db.Posts.Find(id);
-            db.Posts.Remove(post);
-            TempData["message"] = "Postarea a fost stearsa!";
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                db.Posts.Remove(post);
+                TempData["message"] = "Postarea a fost stearsa!";
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra uneri postari care nu va apartine";
+                return RedirectToAction("Index");
+            }
         }
     }
 }
